Sample new batch for generator loss
Fixes a small bug in which the generator loss was caluculated
using the last batch from the critic updates.

NUM_SAMPLES in the data loader is set to this specific number so
there end up being 101 batches which is equal to 100 descriminator
updates + 1 generator update. If it is 10000 like before we run
out of batches before doing a generator update.